doPut and doDelete methods resolved

+++ src/main/java/data/Movie.java
@@ -2,16 +2,16 @@ package data;
 
 public class Movie {
     private String title;
-    private double rating;
+    private Double rating;
     private String poster;
-    private long year;
+    private Integer year;
     private String genre;
     private String director;
     private String plot;
     private String actors;
     private long id;
 
-    public Movie(String title, double rating, String poster, long year, String genre, String director, String plot, String actors, long id) {
+    public Movie(String title, Double rating, String poster, Integer year, String genre, String director, String plot, String actors, long id) {
         this.title = title;
         this.rating = rating;
         this.poster = poster;
@@ -34,11 +34,11 @@ public class Movie {
         this.title = title;
     }
 
-    public double getRating() {
+    public Double getRating() {
         return rating;
     }
 
-    public void setRating(double rating) {
+    public void setRating(Double rating) {
         this.rating = rating;
     }
 
@@ -50,11 +50,11 @@ public class Movie {
         this.poster = poster;
     }
 
-    public long getYear() {
+    public Integer getYear() {
         return year;
     }
 
-    public void setYear(long year) {
+    public void setYear(Integer year) {
         this.year = year;
     }
